Logging changes CR changes
