+++ owncloudComLibrary/src/main/java/com/owncloud/android/lib/common/http/logging/LogRequest.kt
@@ -1,0 +1,11 @@
+package com.owncloud.android.lib.common.http.logging
+
+data class LogRequest(
+    val requestId: String,
+    val method: String,
+    val url: String,
+    val headers: Map<String, String>,
+    val bodyLength: Long,
+    val bodyContentType: String,
+    val body: String,
+)
