+++ owncloudData/build.gradle
@@ -70,7 +70,7 @@ dependencies {
     implementation libs.moshi.kotlin
     ksp libs.moshi.kotlin.codegen
 
-    //Timber
+    // Timber
     implementation libs.timber
 
     // Dependencies for unit tests
